ibusCandidatePopup: Make pagination buttons clickable without stealing focus
The default StButton event handlers will create a temporary grab that will
steal the keyboard focus out of the client, and cause it to pop down the
candidates window.

Avoid the default grabbing behavior, so that focus stays on the Wayland
client while interacting with the ibusCandidatesPopup with the pointer.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/2244
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2497>