windowPreview: Handle window preview activaton differently
Currently, to coordinate the window preview activation on click
and the window preview DnD, there is a ClutterClickAction that
indirectly handles both (the latter through long-press cancellation).

Since there are other means to cancel a ClutterClickAction
other than moving past the drag threshold (say, other touchpoints
appearing), interpreting CLUTTER_LONG_PRESS_CANCEL as "dragged
far enough to initiate window DnD" does not quite hold.

But we cannot (currently) make the actor both draggable and add
a ClutterClickAction for activation and have each handle their
own piece of interaction. Prefer that the draggable implementation
handles DnD entirely on its own, and avoid the ClutterClickAction
with simpler event handlers.

The actions in these event handlers will not get to run if DnD was
initiated, and when they get to happen it will mean DnD did not
get to initiate, which sounds like the desired behavior.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/4586