Connect to DnD signals using connectObject()
Currently DnD uses `disconnectAll()` in a `::destroy` signal handler for
the drag actor. This is a little weird, as usually the connecting object
is in charge of disconnecting as well.

Since we have some nice convenience API for this nowadays, use
`connectObject()` when connecting to DnD signals in the various draggable
actors that we have, which ensures the signal is disconnected as soon as
the draggable actors get destroyed.

This allows removing the call to `disconnectAll()` in DnD.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3796>