Revert "status/volume: Clear device menus when unsetting stream"
Device additions/removals are tracked by GvcMixerControl, which
doesn't change when unsetting the stream. So clearing the menu
manually was a workaround, not a fix.

It's also worth noting that I failed to reproduce the original
issue again, so it's possible that we were working around a
pipewire bug that has since been fixed.

This reverts commit 1b62b7ea0a8aff38ef8c6873cb8b7651abe24044.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2616>