extension-tool: Set reminder for removing old compat code
GSettings schemas are now compiled at install time, so it is no
longer necessary to include the compiled schema in the archive.

However the `gnome-extensions pack` command hasn't been adjusted,
so that it can still be used to produce valid archives for all
supported versions.

To not let that code linger forever, error out when building
a version where GNOME 44 is the oldest supported release.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2639>